requirements.txt
pillow
pytesseract
matplotlib
numpy
sentry-sdk>=0.14.3
